requirements.txt
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
streamlit==1.17.0
